+++ Modules/Intent.Modules.Modelers.Services.DomainInteractions/release-notes.md
@@ -1,3 +1,8 @@
+### Version 2.0.4
+
+- Fixed: Creating CRUD operations for Nested Entities can now be targeted in a specific folder and no longer contains the parent element name in the operation/element names.
+- Improvement: New CRUD Creation screen added.
+
 ### Version 2.0.3
 
 - Improvement: The full path of the class is now displayed on the `Create CRUD Service` and `Create CRUD CQRS Operations` class selection dropdown.